Core Fix, Massive Speed Increase.
Releasing my PointList class. I improve the draw speed significantly of the core type, allowing the project to ultimately be successful. Without this change Android would not be able to rerender fast enough to draw much of anything. I also added in a basic form of how the Matrix for view and Touch work. Also a core touch event, that simply adds lines.